Renamed namespaces.
